# perf: use `willReadFrequently`