This improves the perf of determining the emoji support level, although to see the difference you have to run Chrome in headful mode (interestingly).

```
second-load

┌─────────────┬───────────┐
│     Browser │ chrome    │
│             │ 127.0.0.0 │
├─────────────┼───────────┤
│ Sample size │ 100       │
└─────────────┴───────────┘

┌─────────────┬─────────────┬────────────┬─────────────────────┬───────────────────┬───────────────────┐
│ Benchmark   │ Version     │ Bytes      │            Avg time │    vs this-change │    vs tip-of-tree │
│             │             │            │                     │                   │       tip-of-tree │
├─────────────┼─────────────┼────────────┼─────────────────────┼───────────────────┼───────────────────┤
│ this-change │ <none>      │ 860.51 KiB │ 183.09ms - 187.15ms │                   │            faster │
│             │             │            │                     │          -        │          9% - 13% │
│             │             │            │                     │                   │ 17.95ms - 26.59ms │
├─────────────┼─────────────┼────────────┼─────────────────────┼───────────────────┼───────────────────┤
│ tip-of-tree │ tip-of-tree │ 860.47 KiB │ 203.58ms - 211.21ms │            slower │                   │
│             │             │            │                     │         10% - 14% │          -        │
│             │             │            │                     │ 17.95ms - 26.59ms │                   │
└─────────────┴─────────────┴────────────┴─────────────────────┴───────────────────┴───────────────────┘


```